requirements.txt
altair==5.2.0
annotated-types==0.7.0
anyio==3.5.0
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.2.0
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
elasticsearch==8.11.1
entrypoints==0.4
Faker==22.1.0
favicon==0.7.0
fonttools==4.47.0
gitdb==4.0.11
GitPython==3.1.41
h11==0.12.0
htbuilder==0.6.2
httpcore==0.15.0
httpx==0.23.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.3
jiter==0.8.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.6.0
numpy==1.26.4
openai==1.65.1
packaging==23.2
pandas==2.1.4
pillow==10.4.0
protobuf==4.25.1
pyarrow==14.0.2
pycountry==23.12.11
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.16.2
six==1.17.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
st-annotated-text==4.0.1
streamlit==1.42.2
streamlit-ace==0.1.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.3.0
streamlit-modal==0.1.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3
tqdm==4.66.1
typing-extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
websocket-client==1.7.0
zipp==3.21.0
